+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.8.1
 
-Future
+23rd November 2021
 
 Log Courier / Log Carver